requirements.txt
langchain==0.3.3
langchain-ollama==0.2.0
matplotlib==3.9.2
ollama==0.3.3
pandas==2.2.3
requests==2.32.3
seaborn==0.13.2
fpdf2==2.8.1
pytest==8.3.3
scikit-learn
polars
pyarrow
lightgbm
MLforecast
catboost
ipykernel